+++ math/minuit2/CMakeLists.txt
@@ -17,10 +17,6 @@ endif(NOT CMAKE_PROJECT_NAME STREQUAL ROOT)
 if(CMAKE_PROJECT_NAME STREQUAL ROOT)
   ROOT_STANDARD_LIBRARY_PACKAGE(Minuit2
     HEADERS
-      Minuit2/ABObj.h
-      Minuit2/ABProd.h
-      Minuit2/ABSum.h
-      Minuit2/ABTypes.h
       Minuit2/AnalyticalGradientCalculator.h
       Minuit2/BFGSErrorUpdator.h
       Minuit2/CombinedMinimizer.h
@@ -48,14 +44,7 @@ if(CMAKE_PROJECT_NAME STREQUAL ROOT)
       Minuit2/GradientCalculator.h
       Minuit2/HessianGradientCalculator.h
       Minuit2/InitialGradientCalculator.h
-      Minuit2/LASymMatrix.h
-      Minuit2/LAVector.h
-      Minuit2/LaInverse.h
-      Minuit2/LaOuterProduct.h
-      Minuit2/LaProd.h
-      Minuit2/LaSum.h
       Minuit2/MPIProcess.h
-      Minuit2/MatrixInverse.h
       Minuit2/MinimumBuilder.h
       Minuit2/MinimumError.h
       Minuit2/MinimumErrorUpdator.h
@@ -117,11 +106,9 @@ if(CMAKE_PROJECT_NAME STREQUAL ROOT)
       Minuit2/SinParameterTransformation.h
       Minuit2/SqrtLowParameterTransformation.h
       Minuit2/SqrtUpParameterTransformation.h
-      Minuit2/StackAllocator.h
       Minuit2/VariableMetricBuilder.h
       Minuit2/VariableMetricEDMEstimator.h
       Minuit2/VariableMetricMinimizer.h
-      Minuit2/VectorOuterProduct.h
       TMinuit2TraceObject.h
     SOURCES
       src/AnalyticalGradientCalculator.cxx
@@ -138,12 +125,6 @@ if(CMAKE_PROJECT_NAME STREQUAL ROOT)
       src/FumiliStandardMaximumLikelihoodFCN.cxx
       src/HessianGradientCalculator.cxx
       src/InitialGradientCalculator.cxx
-      src/LaEigenValues.cxx
-      src/LaInnerProduct.cxx
-      src/LaInverse.cxx
-      src/LaOuterProduct.cxx
-      src/LaSumOfElements.cxx
-      src/LaVtMVSimilarity.cxx
       src/MPIProcess.cxx
       src/MinimumBuilder.cxx
       src/Minuit2Minimizer.cxx
@@ -158,6 +139,7 @@ if(CMAKE_PROJECT_NAME STREQUAL ROOT)
       src/MnHesse.cxx
       src/MnLineSearch.cxx
       src/MnMachinePrecision.cxx
+      src/MnMatrix.cxx
       src/MnMinos.cxx
       src/MnParabolaFactory.cxx
       src/MnParameterScan.cxx
@@ -190,17 +172,7 @@ if(CMAKE_PROJECT_NAME STREQUAL ROOT)
       src/VariableMetricBuilder.cxx
       src/VariableMetricEDMEstimator.cxx
       src/mnbins.cxx
-      src/mndasum.cxx
-      src/mndaxpy.cxx
-      src/mnddot.cxx
-      src/mndscal.cxx
-      src/mndspmv.cxx
-      src/mndspr.cxx
-      src/mnlsame.cxx
-      src/mnteigen.cxx
       src/mntplot.cxx
-      src/mnvert.cxx
-      src/mnxerbla.cxx
     DICTIONARY_OPTIONS
       -writeEmptyRootPCM
     DEPENDENCIES
