+++ math/minuit2/inc/Minuit2/MnPrint.h
@@ -31,12 +31,6 @@ std::ostream &operator<<(std::ostream &, const FunctionMinimum &);
 class MinimumState;
 std::ostream &operator<<(std::ostream &, const MinimumState &);
 
-class LAVector;
-std::ostream &operator<<(std::ostream &, const LAVector &);
-
-class LASymMatrix;
-std::ostream &operator<<(std::ostream &, const LASymMatrix &);
-
 class MnUserParameters;
 std::ostream &operator<<(std::ostream &, const MnUserParameters &);
 
@@ -110,14 +104,6 @@ public:
 
    static void AddFilter(const char *prefix);
    static void ClearFilter();
-
-   // Whether to cut the maximum number of parameters shown for vector and matrices
-   // set maximum number of printed parameters and return previous value
-   // A negative value will mean all parameters are printed
-   static int SetMaxNP(int value);
-
-   // retrieve maximum number of printed parameters
-   static int MaxNP();
 
    // set print level and return the previous one
    int SetLevel(int level);
