+++ math/minuit2/src/CMakeLists.txt
@@ -7,10 +7,6 @@
 add_subdirectory(math)
 
 set(MINUIT2_HEADERS
-    ABObj.h
-    ABProd.h
-    ABSum.h
-    ABTypes.h
     AnalyticalGradientCalculator.h
     BFGSErrorUpdator.h
     CombinedMinimizer.h
@@ -38,14 +34,7 @@ set(MINUIT2_HEADERS
     GradientCalculator.h
     HessianGradientCalculator.h
     InitialGradientCalculator.h
-    LASymMatrix.h
-    LAVector.h
-    LaInverse.h
-    LaOuterProduct.h
-    LaProd.h
-    LaSum.h
     MPIProcess.h
-    MatrixInverse.h
     MinimumBuilder.h
     MinimumError.h
     MinimumErrorUpdator.h
@@ -69,6 +58,7 @@ set(MINUIT2_HEADERS
     MnHesse.h
     MnLineSearch.h
     MnMachinePrecision.h
+    MnMatrix.h
     MnMatrix.h
     MnMatrixfwd.h
     MnMigrad.h
@@ -106,11 +96,9 @@ set(MINUIT2_HEADERS
     SinParameterTransformation.h
     SqrtLowParameterTransformation.h
     SqrtUpParameterTransformation.h
-    StackAllocator.h
     VariableMetricBuilder.h
     VariableMetricEDMEstimator.h
     VariableMetricMinimizer.h
-    VectorOuterProduct.h
     )
 
 set(MINUIT2_SOURCES
@@ -127,12 +115,6 @@ set(MINUIT2_SOURCES
     FumiliStandardMaximumLikelihoodFCN.cxx
     HessianGradientCalculator.cxx
     InitialGradientCalculator.cxx
-    LaEigenValues.cxx
-    LaInnerProduct.cxx
-    LaInverse.cxx
-    LaOuterProduct.cxx
-    LaSumOfElements.cxx
-    LaVtMVSimilarity.cxx
     MPIProcess.cxx
     MinimumBuilder.cxx
     Minuit2Minimizer.cxx
@@ -177,17 +159,7 @@ set(MINUIT2_SOURCES
     VariableMetricBuilder.cxx
     VariableMetricEDMEstimator.cxx
     mnbins.cxx
-    mndasum.cxx
-    mndaxpy.cxx
-    mnddot.cxx
-    mndscal.cxx
-    mndspmv.cxx
-    mndspr.cxx
-    mnlsame.cxx
-    mnteigen.cxx
     mntplot.cxx
-    mnvert.cxx
-    mnxerbla.cxx
     )
 
 prepend_path(MINUIT2_HEADERS "${Minuit2_SOURCE_DIR}/inc/Minuit2" ${MINUIT2_HEADERS})
