+++ math/minuit2/src/FumiliErrorUpdator.cxx
@@ -17,7 +17,6 @@
 #include "Minuit2/MnMatrix.h"
 #include "Minuit2/MinimumError.h"
 #include "Minuit2/MinimumState.h"
-#include "Minuit2/LaSum.h"
 #include "Minuit2/MnPrint.h"
 
 #include <limits>
@@ -25,8 +24,6 @@
 namespace ROOT {
 
 namespace Minuit2 {
-
-double sum_of_elements(const LASymMatrix &);
 
 MinimumError
 FumiliErrorUpdator::Update(const MinimumState &s0, const MinimumParameters &p1, const FunctionGradient &g1) const
