+++ math/minuit2/src/MnPosDef.cxx
@@ -18,8 +18,6 @@ namespace ROOT {
 
 namespace Minuit2 {
 
-LAVector eigenvalues(const LASymMatrix &);
-
 MinimumState MnPosDef::operator()(const MinimumState &st, const MnMachinePrecision &prec) const
 {
    // interface from minimum state
